refactor : 불필요한 렌더링을 막기 위해 코드를 분리하였지만 아직 문제를 개선하지 못함함
