+++ frontend/src/pages/my-page/api/purchaseDuck.ts
@@ -1,12 +1,24 @@
-async function purchaseDuck() {
-  try {
-    const response = await fetch("/api/users/purchaseduck");
-    if (!response.ok) {
-      throw new Error("오리를 구매하는데 실패했습니다.");
-    }
-  } catch (error) {
-    console.error(error);
+import { QueryClient } from "@tanstack/react-query";
+import { userInfoQueries } from "@/shared/lib/auth/authQuery";
+import { updateQueryClient } from "@/shared/lib/updateQueryClient";
+
+async function purchaseDuck(currentDuck: number, queryClient: QueryClient) {
+  if (currentDuck < 30) {
+    return;
   }
+
+  const response = await fetch("/api/users/purchaseduck", {
+    cache: "no-cache",
+  });
+  if (!response.ok) {
+    console.error("Failed to purchase duck");
+    return;
+  }
+  updateQueryClient(queryClient, userInfoQueries.queryKey, (prev) => ({
+    ...prev,
+    duck: prev.userInfo.duck - 30,
+    realDuck: prev.userInfo.realDuck + 1,
+  }));
 }
 
 export { purchaseDuck };
