+++ frontend/src/pages/my-page/ui/FallingDuck.tsx
@@ -30,4 +30,4 @@ const FallingDuck = memo(() => {
   );
 });
 
-export { FallingDuck };
+export default FallingDuck;
